requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
urllib3==2.0.7
PyMuPDF==1.25.5

markdownify==1.1.0
pillow==11.1.0
pytesseract==0.3.13
paddlepaddle==3.0.0
paddleocr==2.10.0
setuptools==78.1.0

tesserocr=2.8.0